DRY out the config syncing.
